tsan_server/requirements.txt
django
graphene-django
sqlite3
